+++ brocooly.config.js
@@ -17,8 +17,6 @@ module.exports = {
 			'./resources/assets/js/app.js',
 			'./resources/assets/sass/app.scss',
 		],
-
-		'comments-reply': './resources/assets/js/components/comments-reply.js',
 	},
 
 	/**