+++ web/app/themes/brocooly/config/app.php
@@ -6,22 +6,10 @@
  * @since 0.1.0
  */
 
-use Theme\Models\WP\Tag;
 use Theme\Models\WP\Page;
 use Theme\Models\WP\Post;
-use Theme\Models\WP\Category;
-
-use Theme\Hooks\GetSearchForm;
-use Theme\Hooks\WpEnqueueScripts;
-use Theme\Hooks\DisableAuthorArchives;
-use Theme\Hooks\LazyLoadContentImages;
-
-use Theme\Http\Middleware\UserLoggedIn;
 
 use Theme\Providers\ThemeServiceProvider;
-use Theme\Providers\GutenbergServiceProvider;
-use Theme\Containers\PageSection\Providers\PageServiceProvider;
-use Theme\Containers\BlogSection\Providers\PostServiceProvider;
 
 return [
 
@@ -40,10 +28,7 @@ return [
 		Page::class,
 	],
 
-	'taxonomies' => [
-		Category::class,
-		Tag::class,
-	],
+	'taxonomies' => [],
 
 	/**
 	 * --------------------------------------------------------------------------
@@ -56,10 +41,7 @@ return [
 	 *
 	 * @since 0.3.0
 	 */
-	'hooks'      => [
-		GetSearchForm::class,
-		WpEnqueueScripts::class,
-	],
+	'hooks'      => [],
 
 	/**
 	 * --------------------------------------------------------------------------
@@ -70,9 +52,6 @@ return [
 	 */
 	'providers'  => [
 		ThemeServiceProvider::class,
-		GutenbergServiceProvider::class,
-		PostServiceProvider::class,
-		PageServiceProvider::class,
 	],
 
 	/**
@@ -84,8 +63,6 @@ return [
 	 *
 	 * @since 0.8.0
 	 */
-	'middleware' => [
-		UserLoggedIn::class,
-	],
+	'middleware' => [],
 
 ];