+++ web/app/themes/brocooly/src/Containers/FrontPageSection/Tasks/GetHelloWorld.php
@@ -4,13 +4,14 @@ declare(strict_types=1);
 
 namespace Theme\Containers\FrontPageSection\Tasks;
 
-class GetMainScreenHelloPhrase
+class GetHelloWorld
 {
 	public function run() {
-		return sprintf(
+		$hello = sprintf(
 			/* translators: 1 - theme version */
 			esc_html__( 'Welcome to Brocooly Framework v%s', 'brocooly' ),
 			wp_get_theme()->get( 'Version' ),
 		);
+		return $hello;
 	}
 }